+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Upcoming
-* Increased Vigour and Soulbound cost
+* Increased Piercing, Vigour and Soulbound cost
 * Fixed supplementaries urn loot table
 * Fixed Burst of Arrows' arrows being pickupable
 * Fixed Grindstone consuming multiple books (if more are placed in the slot the recipe will not work)